don't import from pages into _app. disable sentry debug. upgrade ory/client
