+++ apps/hash-frontend/src/shared/get-block-dom-id.ts
@@ -1,0 +1,2 @@
+export const getBlockDomId = (blockEntityId: string) =>
+  `entity-${blockEntityId}`;
